+++ web/src/narrow_banner.ts
@@ -1,15 +1,18 @@
 import $ from "jquery";
 import assert from "minimalistic-assert";
 
+import {all_messages_data} from "./all_messages_data";
 import {$t, $t_html} from "./i18n";
 import type {NarrowBannerData, SearchData} from "./narrow_error";
 import {narrow_error} from "./narrow_error";
 import * as narrow_state from "./narrow_state";
 import {page_params} from "./page_params";
 import * as people from "./people";
+import * as settings_config from "./settings_config";
 import * as spectators from "./spectators";
 import {realm} from "./state_data";
 import * as stream_data from "./stream_data";
+import {get_user_group_from_id} from "./user_groups";
 
 const SPECTATOR_STREAM_NARROW_BANNER = {
     title: "",
@@ -69,7 +72,7 @@ function retrieve_search_query_data(): SearchData {
     return search_string_result;
 }
 
-function pick_empty_narrow_banner(): NarrowBannerData {
+export function pick_empty_narrow_banner(): NarrowBannerData {
     const default_banner = {
         title: $t({defaultMessage: "There are no messages here."}),
         // Spectators cannot start a conversation.
@@ -298,6 +301,45 @@ function pick_empty_narrow_banner(): NarrowBannerData {
             }
             const user_ids = people.emails_strings_to_user_ids_array(first_operand);
             assert(user_ids !== undefined);
+            const user_ids_string = user_ids.join(",");
+            if (!people.user_can_direct_message(user_ids_string)) {
+                const {name} = get_user_group_from_id(realm.realm_direct_message_permission_group);
+                if (name === "role:nobody") {
+                    return {
+                        title: $t({
+                            defaultMessage:
+                                "You are not allowed to send direct messages in this organization.",
+                        }),
+                    };
+                }
+                const display_name = settings_config.system_user_groups_list.find(
+                    (group) => group.name === name,
+                )?.display_name;
+                assert(display_name !== undefined);
+                return {
+                    title: $t(
+                        {
+                            defaultMessage:
+                                "{allowed_user_group} must be in every direct message conversation.",
+                        },
+                        {allowed_user_group: display_name.replace(" and ", " or ")},
+                    ),
+                };
+            }
+            const previous_messages_exist = all_messages_data
+                .all_messages()
+                .find((message) => message.is_private && message.to_user_ids === user_ids_string);
+            if (
+                !previous_messages_exist &&
+                !people.user_can_initiate_direct_message_thread(user_ids_string)
+            ) {
+                return {
+                    title: $t({
+                        defaultMessage:
+                            "You are not allowed to start direct message conversations in this organization.",
+                    }),
+                };
+            }
             if (!first_operand.includes(",")) {
                 // You have no direct messages with this person
                 if (people.is_current_user(first_operand)) {
@@ -369,6 +411,45 @@ function pick_empty_narrow_banner(): NarrowBannerData {
             if (!person_in_dms) {
                 return {
                     title: $t({defaultMessage: "This user does not exist!"}),
+                };
+            }
+            const person_id_string = person_in_dms.user_id.toString();
+            if (!people.user_can_direct_message(person_id_string)) {
+                const {name} = get_user_group_from_id(realm.realm_direct_message_permission_group);
+                if (name === "role:nobody") {
+                    return {
+                        title: $t({
+                            defaultMessage:
+                                "You are not allowed to send direct messages in this organization.",
+                        }),
+                    };
+                }
+                const display_name = settings_config.system_user_groups_list.find(
+                    (group) => group.name === name,
+                )?.display_name;
+                assert(display_name !== undefined);
+                return {
+                    title: $t(
+                        {
+                            defaultMessage:
+                                "{allowed_user_group} must be in every direct message conversation.",
+                        },
+                        {allowed_user_group: display_name.replace(" and ", " or ")},
+                    ),
+                };
+            }
+            const previous_messages_exist = all_messages_data
+                .all_messages()
+                .find((message) => message.is_private && message.to_user_ids === person_id_string);
+            if (
+                !previous_messages_exist &&
+                !people.user_can_initiate_direct_message_thread(person_id_string)
+            ) {
+                return {
+                    title: $t({
+                        defaultMessage:
+                            "You are not allowed to start direct message conversations in this organization.",
+                    }),
                 };
             }
             if (people.is_current_user(first_operand)) {